// src/api/generated/models/meetingResponseDtoCurrentPhase.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Meetings Quality API
 * API для платформы отслеживания качества встреч
 * OpenAPI spec version: 1.0
 */

export type MeetingResponseDtoCurrentPhase =
  typeof MeetingResponseDtoCurrentPhase[keyof typeof MeetingResponseDtoCurrentPhase];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MeetingResponseDtoCurrentPhase = {
  discussion: "discussion",
  evaluation: "evaluation",
  summary: "summary",
  finished: "finished",
} as const;