// src/api/generated/models/changePhaseDtoPhase.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Meetings Quality API
 * API для платформы отслеживания качества встреч
 * OpenAPI spec version: 1.0
 */

/**
 * Новая фаза встречи
 */
export type ChangePhaseDtoPhase =
  typeof ChangePhaseDtoPhase[keyof typeof ChangePhaseDtoPhase];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ChangePhaseDtoPhase = {
  discussion: "discussion",
  evaluation: "evaluation",
  summary: "summary",
  finished: "finished",
} as const;